V3

+++ ft_printf.h
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef  FT_PRINTF_H
+#ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
 # include <stdarg.h>
@@ -20,16 +20,15 @@
 size_t	ft_strlen(const char *str);
 void	ft_putchar_fd(char s, int fd);
 void	ft_itoa_dec(int n, char **str);
-void	ft_lower_hex(unsigned long long n, char **str);
+void	ft_lower_hex(unsigned int n, char **str);
 void	ft_pointer_hex(unsigned long long n, char **str);
-void	ft_upper_hex(unsigned long long n, char **str);
-void    ft_itoa_unsign(unsigned int n, char **str);
+void	ft_upper_hex(unsigned int n, char **str);
+void	ft_itoa_unsign(unsigned int n, char **str);
 void	ft_char(int a, char **str);
 void	ft_char_word(int a, char **str);
 void	ft_string(char *src, char **ptr);
 void	ft_putstr_fd(char *s, int fd);
 char	*ft_strdup(char *src);
 size_t	ft_printf(const char *word, ...);
-
 
 #endif
